+++ crates/ic-query/src/nns/macros/mod.rs
@@ -1,4 +1,6 @@
 #[macro_use]
+mod leaf_refresh;
+#[macro_use]
 mod leaf_requests;
 #[macro_use]
 mod network;